# modules/backend/swagger/swagger.yaml
basePath: /api/v1
definitions:
  SurveyListResponse:
    properties:
      limit:
        type: integer
      list:
        items:
          $ref: '#/definitions/entity.ViewSurveyList'
        type: array
      page:
        type: integer
    type: object
  SurveyResponseCompleteRequest:
    properties:
      answers:
        additionalProperties: true
        type: object
      responseID:
        type: string
      slug:
        type: string
    required:
    - answers
    - responseID
    - slug
    type: object
  entity.Answers:
    additionalProperties: true
    type: object
  entity.Conditional:
    properties:
      questionId:
        type: string
      value:
        type: string
    type: object
  entity.Option:
    properties:
      id:
        type: string
      value:
        type: string
    type: object
  entity.Question:
    properties:
      conditional:
        $ref: '#/definitions/entity.Conditional'
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      order:
        type: integer
      placeholder:
        type: string
      required:
        type: boolean
      text:
        type: string
      type:
        $ref: '#/definitions/entity.QuestionType'
    type: object
  entity.QuestionType:
    enum:
    - text
    - number
    - select
    - multi-select
    - boolean
    type: string
    x-enum-varnames:
    - QuestionTypeText
    - QuestionTypeNumber
    - QuestionTypeSelect
    - QuestionTypeMultiSelect
    - QuestionTypeBoolean
  entity.Survey:
    properties:
      createdAt:
        type: string
      createdBy:
        description: Optional creator name
        type: string
      description:
        type: string
      finishesAt:
        type: string
      id:
        type: string
      minCompletionTimeMin:
        description: Minimum completion time in minutes
        type: integer
      questions:
        items:
          $ref: '#/definitions/entity.Question'
        type: array
      slug:
        type: string
      tags:
        description: Optional tags for categorization
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  entity.ViewSurveyList:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      finishesAt:
        type: string
      id:
        type: string
      minCompletionTimeMin:
        type: integer
      participants:
        description: Number of completed responses
        type: integer
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  entity.ViewSurveyResponse:
    properties:
      answers:
        $ref: '#/definitions/entity.Answers'
      id:
        type: string
      startedAt:
        type: string
      surveySlug:
        type: string
    type: object
  handler.UserProfile:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      picture:
        type: string
    type: object
host: localhost:4004
info:
  contact:
    email: support@maasanketi.co
    name: API Support
    url: http://www.maasanketi.co/support
  description: API for the Maasanketi platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Maasanketi API
  version: "1.0"
paths:
  /api/v1/user/profile:
    get:
      consumes:
      - application/json
      description: View the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserProfile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: View user profile
      tags:
      - user
  /healthcheck:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Show the status of server.
      tags:
      - HealthCheck
  /surveys:
    get:
      consumes:
      - application/json
      description: List all available surveys with pagination, filtering, and sorting
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Sort field (created_at_asc, created_at_desc, most_participants,
          finishes_at_asc, finishes_at_desc)
        in: query
        name: sort
        type: string
      - description: Hide expired surveys
        in: query
        name: hideExpired
        type: boolean
      - description: Text search query for title and description
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SurveyListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all surveys with pagination and filtering
      tags:
      - surveys
  /surveys/{slug}:
    get:
      consumes:
      - application/json
      description: Get a survey by its Slug
      parameters:
      - description: Survey Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Survey'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a survey by Slug
      tags:
      - surveys
  /surveys/{slug}/responses:
    post:
      consumes:
      - application/json
      description: Start a new response for a survey
      parameters:
      - description: Survey Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Is Anonymous
        in: body
        name: isAnonymous
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.ViewSurveyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Start a new survey response
      tags:
      - survey-responses
  /surveys/{slug}/responses/{responseId}:
    patch:
      consumes:
      - application/json
      description: Complete a survey response with the provided data
      parameters:
      - description: Survey Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Response ID
        in: path
        name: responseId
        required: true
        type: string
      - description: Response data
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/SurveyResponseCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ViewSurveyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Complete a survey response
      tags:
      - survey-responses
  /surveys/{slug}/stats:
    get:
      consumes:
      - application/json
      description: Get aggregated statistics for a specific survey
      parameters:
      - description: Survey Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get statistics for a survey
      tags:
      - surveys
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the ID token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"